refactor(web): migrate months to database queries
